+++ src/components/Footer/JoinTheCommunity.jsx
